refactor(web page reader): change plugin name, add request params schema to the OpenAPI doc
