+++ src/api-docs/openAPIDocumentGenerator.ts
@@ -1,8 +1,8 @@
 import { OpenApiGeneratorV3, OpenAPIRegistry } from '@asteasolutions/zod-to-openapi';
 
-import { articleReaderRegistry } from '@/routes/articleReader/articleReaderRouter';
 import { healthCheckRegistry } from '@/routes/healthCheck/healthCheckRouter';
 import { powerpointGeneratorRegistry } from '@/routes/powerpointGenerator/powerpointGeneratorRouter';
+import { articleReaderRegistry } from '@/routes/webPageReader/webPageReaderRouter';
 import { wordGeneratorRegistry } from '@/routes/wordGenerator/wordGeneratorRouter';
 import { transcriptRegistry } from '@/routes/youtubeTranscript/transcriptRouter';
